refactor: Change distance data type from Int to Double
Increase the precision of distance tracking by changing the data type for `distanceInMeters` and related calculations from `Int` to `Double`. This refactoring ensures more accurate distance measurements and statistics throughout the application.

This change propagates through all layers of the architecture:

-   **Data Layer:**
    -   Updated `RunEntity` and `RunDao` to store and query distance as a `Double` in the database.
    -   Modified `RunRepositoryImpl` and `MockRunRepositoryImpl` to handle `Double` for total distance calculations.

-   **Domain Layer:**
    -   Changed `Run` model and `RunRepository` interface to use `Double` for `distanceInMeters`.
    -   Adjusted `GetTotalStatsUseCase` to return total distance as a `Flow<Double>`.

-   **Presentation Layer & Service:**
    -   Updated `TrackingState`, `MainState`, and `TrackingService` to manage distance as a `Double`.
    -   Modified `TrackingCalculator` to accept `Double` for average speed calculation, improving precision.
    -   Adjusted UI components in `MainScreen` to format the distance `Double` to two decimal places (e.g., "5.25 km").

+++ app/src/main/java/com/survivalcoding/runningtracker/core/util/TrackingCalculator.kt
@@ -23,10 +23,10 @@ object TrackingCalculator {
     /**
      * Calculates the average speed in km/h.
      */
-    fun calculateAvgSpeed(distanceInMeters: Int, timeInMillis: Long): Float {
+    fun calculateAvgSpeed(distanceInMeters: Double, timeInMillis: Long): Float {
         val timeInSeconds = timeInMillis / 1000f
         return if (timeInSeconds > 0) {
-            (distanceInMeters / timeInSeconds) * 3.6f
+            ((distanceInMeters / timeInSeconds) * 3.6).toFloat()
         } else {
             0f
         }
